refactor(web): rename lib/artifactSource → lib/artifactSourceLabels
Disambiguates from the backend artifact-source/ directory. The file
contains only UI label functions (sourceLabel, sourceBadgeLabel), not
artifact source infrastructure.
